:recycle: Reorganiza middlewares
