+++ api/routes.go
@@ -18,7 +18,7 @@ func V1(router *gin.Engine) {
 	v1 := router.Group("v1")
 	v1.Use(timingMetrics())
 	v1.Use(returnHeaders())
-	v1.POST("/boleto/register", Authentication, ParseBoleto, ValidateRegisterV1, Logger, registerBoleto)
+	v1.POST("/boleto/register", authentication, parseBoleto, validateRegisterV1, logger, registerBoleto)
 	v1.GET("/boleto/:id", getBoletoByID)
 }
 
@@ -27,5 +27,5 @@ func V2(router *gin.Engine) {
 	v2 := router.Group("v2")
 	v2.Use(timingMetrics())
 	v2.Use(returnHeaders())
-	v2.POST("/boleto/register", Authentication, ParseBoleto, ValidateRegisterV2, Logger, registerBoleto)
+	v2.POST("/boleto/register", authentication, parseBoleto, validateRegisterV2, logger, registerBoleto)
 }
